+++ src/components/upload_stepper/UploadResultsContent.js
@@ -6,14 +6,14 @@ import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
 import {useDispatch, useSelector} from "react-redux";
-import LeftSpinner from "../spinners/LeftSpinner";
+import LeftSpinner from "../widgets/spinners/LeftSpinner";
 import CheckCircleOutlineIcon from '@material-ui/icons/CheckCircleOutline';
 import ErrorOutlineIcon from '@material-ui/icons/ErrorOutline';
 import { jsx } from '@emotion/core';
-import {actionCreators} from "../../../redux/actionCreators";
-import colors from "../../../styles/colors";
-import RetryUploadButton from "../buttons/RetryButtons/RetryUploadButton";
-import RetryStartUploadOverButton from "../buttons/RetryButtons/RetryStartUploadOverButton";
+import {actionCreators} from "../../redux/actionCreators";
+import colors from "../../styles/colors";
+import RetryUploadButton from "../widgets/buttons/RetryButtons/RetryUploadButton";
+import RetryStartUploadOverButton from "../widgets/buttons/RetryButtons/RetryStartUploadOverButton";
 
 /**
  * This component watches for the upload state to change and then renders the appropriate
