+++ src/components/upload_stepper/UploadStepConnector.js
@@ -1,7 +1,7 @@
 import React from "react";
 import withStyles from "@material-ui/core/styles/withStyles";
 import StepConnector from "@material-ui/core/StepConnector";
-import colors from "../../../styles/colors";
+import colors from "../../styles/colors";
 
 const CustomConnector = withStyles({
   active: {
